# Add DataFusion datasource implementation in Python for pandas and DataFrame Interchange