--- a/python/vegafusion/vegafusion/datasource/__init__.py
+++ b/python/vegafusion/vegafusion/datasource/__init__.py
@@ -1,0 +1,2 @@
+from .dfi_datasource import DfiDatasource
+from .pandas_datasource import PandasDatasource