--- a/python/vegafusion/vegafusion/datasource/datasource.py
+++ b/python/vegafusion/vegafusion/datasource/datasource.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+from typing import Iterable
+import pyarrow as pa
+
+
+class Datasource(ABC):
+    @abstractmethod
+    def schema(self) -> pa.Schema:
+        raise NotImplementedError()
+
+    @abstractmethod
+    def fetch(self, columns: Iterable[str]) -> pa.Table:
+        raise NotImplementedError()